@@ -14,21 +14,23 @@ description: Connect and get accurate location data from Bluetooth GPS, Galileo,
 version: 1.0.43+43
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.19.6 <3.0.0'  # dart ver
+  flutter: 3.10.2  # flutter ver
+
 
 dependencies:
   flutter:
     sdk: flutter
 
-  flutter_markdown: 0.5.2
+  flutter_markdown: 0.6.15
   share: 2.0.1
   package_info: 2.0.0
-  progress_dialog: 1.2.4
   flutter_gallery_assets: 1.0.2
-  preferences: 5.2.1
+
   url_launcher: 6.0.3
   wakelock: 0.5.6
   shared_preferences:
+  pref: ^2.7.1
 
 
 dev_dependencies:
